added support for restoring deleted records
the hope was to retain the record's original id, but this
seems to require unacceptable hackery, so living with getting
a new id for now.

left in the best hack we came up with as an example.

this commit also fixes up some of the tests following the backend
refactor.