Removed CollectionManagement from registered models
They don't need their own history tracking because they are always
nested under other records.

And, because their uris have repo_ids but they don't have a repo_id,
they were causing errors to be thrown when their parent records
were being saved -- they are deleted and recreated (it's a nested
record thing), so history dutifully tried to record the fact of the
deletion but hasn't got the foggiest what the repo_id is.

Successfully mixed tense and number. grammer for the win.

... world saved.