Make version numbers sequential
An earlier change stopped versions being written when a `lock_version`
is bumped on a record with no associated changes to its data.

This led to gaps in the list of version numbers in the record's history.

The fix here is to introduce a new `revision` column that provides a
sequential version number starting at 1.