When restoring, update the lock_version for all subrecords that have uris
There are some kinds of subrecords that are also top-level records. These
have their own uri and lock_version. So, when restoring, it is not sufficient
to update the lock_version only at the top-level. Rather, it is necessary
to walk down the record structure looking for subrecords with uris and updating
their lock_versions too.