When diffing, ensure out[a_key] is an array when it should be.
It looks like it could be possible for out[a_key] to contain a diff
hash when there are significant schame changes between the versions
being diffed. This is a pretty obscure edge case, and not yet confirmed,
but this change is sound regardless and should ensure we don't end up
trying to push to a hash.

This commit also contains a few small deletions of no-op lines discovered
during analysis.