Ensure that a failure to diff doesn't blow up a request.
If the diff fails then return the clean json and log the exception.

Of course, the diff shouldn't fail, but if there have been schema changes
between the versions being compared then the diff routine might get confused.
In these circumstances, it is better to fail gracefully and give the caller
the json.